Stop removing temporary directory on managed hosts
A temporary directory is created by the preflight and software
roles. The removal of the temporary directory in software role
was commented out by Mark Stillings. Now I comment out the
removal in preflight role as well.

It is less dangerous to leave our files on the managed hosts
than removing directories.